Added more routes to Auth middleware, provided basic authorization

+++ app/Http/Controllers/CommentsController.php
@@ -28,7 +28,7 @@ class CommentsController extends Controller
     public function store(Request $request)
     {
         $comment = new \App\Comment;
-        $comment->user_id = \Auth::user->$id;
+        $comment->user_id = \Auth::user()->$id;
         $comment->comment_id = $request->comment_id;
         $comment->post_id = $request->post_id;
         $comment->content = $request->comment_content;
@@ -58,8 +58,12 @@ class CommentsController extends Controller
     public function update(Request $request, $id)
     {
         $comment = \App\Comment::find($id);
-        $comment->content = $request->comment_content;
-        $comment->save();
+        if ($comment->user_id == \Auth::user()->id) {
+            $comment->content = $request->comment_content;
+            $comment->save();
+        else {
+            return response("Unauthorized", 403);
+        }
 
         return $comment;
     }
@@ -73,7 +77,11 @@ class CommentsController extends Controller
     public function destroy($id)
     {
         $comment = \App\Comment::find($id);
-        $comment->delete();
+        if ($comment->user_id == \Auth::user()->id) {
+            $comment->delete();
+        else {
+            return response("Unauthorized", 403);
+        }
         return $comment;
     }
 }
