+++ app/Http/Controllers/PostsController.php
@@ -61,11 +61,15 @@ class PostsController extends Controller
     public function update(Request $request, $id)
     {
         $post = \App\Post::find($id);
-        $post->title = $request->title;
-        $post->content = $request->post_content;
-        $post->url = $request->url;
+        if ($post->user_id == \Auth::user()->id) {
+            $post->title = $request->title;
+            $post->content = $request->post_content;
+            $post->url = $request->url;
 
-        $post->save();
+            $post->save();
+        else {
+            return response("Unauthorized", 403);
+        }
 
         return $post;
     }
@@ -79,7 +83,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = \App\Post::find($id);
-        $post->delete();
+        if ($post->user_id == \Auth::user()->id) {
+            $post->delete();
+        else {
+            return response("Unauthorized", 403);
+        }
         return $post;
     }
 }
