+++ app/Http/Controllers/SubbredditsController.php
@@ -60,9 +60,13 @@ class SubbredditsController extends Controller
     public function update(Request $request, $id)
     {
         $subbreddit = \App\Subbreddit::find($id);
-        $subbreddit->name = $request->name;
-        $subbreddit->description = $request->description;
-        $subbreddit->save();
+        if ($subbreddit->user_id == \Auth::user()->id) {
+            $subbreddit->name = $request->name;
+            $subbreddit->description = $request->description;
+            $subbreddit->save();
+        } else {
+            return response("Unauthorized", 403);
+        }
 
         return $subbreddit;
     }
@@ -76,7 +80,11 @@ class SubbredditsController extends Controller
     public function destroy($id)
     {
         $subbreddit = \App\Subbreddit::find($id);
-        $subbreddit->delete();
+        if ($subbreddit->user_id == \Auth::user()->id) {
+            $subbreddit->delete();
+        } else {
+            return response("Unauthorized", 403);
+        }
         return $subbreddit;
     }
 }
